fix padentic

+++ src/api/nlopt-in.hpp
@@ -361,7 +361,7 @@ namespace nlopt {
       if (!d) throw std::bad_alloc();
       d->o = this; d->mf = mf; d->f_data = f_data; d->f = NULL; d->vf = NULL;
       d->munge_destroy = d->munge_copy = NULL;
-      mythrow(nlopt_add_inequality_mconstraint(o, tol.size(), mymfunc, d,
+      mythrow(nlopt_add_inequality_mconstraint(o, (unsigned)tol.size(), mymfunc, d,
 					       tol.empty() ? NULL : &tol[0]));
     }
 
@@ -390,7 +390,7 @@ namespace nlopt {
       if (!d) throw std::bad_alloc();
       d->o = this; d->mf = mf; d->f_data = f_data; d->f = NULL; d->vf = NULL;
       d->munge_destroy = d->munge_copy = NULL;
-      mythrow(nlopt_add_equality_mconstraint(o, tol.size(), mymfunc, d,
+      mythrow(nlopt_add_equality_mconstraint(o, (unsigned)tol.size(), mymfunc, d,
 					     tol.empty() ? NULL : &tol[0]));
     }
 
@@ -420,7 +420,7 @@ namespace nlopt {
       if (!d) throw std::bad_alloc();
       d->o = this; d->mf = mf; d->f_data = f_data; d->f = NULL; d->vf = NULL;
       d->munge_destroy = md; d->munge_copy = mc;
-      mythrow(nlopt_add_inequality_mconstraint(o, tol.size(), mymfunc, d,
+      mythrow(nlopt_add_inequality_mconstraint(o, (unsigned)tol.size(), mymfunc, d,
 					       tol.empty() ? NULL : &tol[0]));
     }
     void add_equality_mconstraint(mfunc mf, void *f_data,
@@ -430,7 +430,7 @@ namespace nlopt {
       if (!d) throw std::bad_alloc();
       d->o = this; d->mf = mf; d->f_data = f_data; d->f = NULL; d->vf = NULL;
       d->munge_destroy = md; d->munge_copy = mc;
-      mythrow(nlopt_add_equality_mconstraint(o, tol.size(), mymfunc, d,
+      mythrow(nlopt_add_equality_mconstraint(o, (unsigned)tol.size(), mymfunc, d,
 					     tol.empty() ? NULL : &tol[0]));
     }
